Vector Fully tested working
Was using a call to VectorReplace() inside VectorInsert() to write
the new element to the vector.  This was a bug because
VectorReplace() calls the FreeFunction on the element at the location
where it is writing.  But VectorInsert() is only moving elements,
not torching them, so freeing the element destroyed a still-active
reference (the next reference over).